Move `BuildMode` to `util.ts`
