+++ lib/analyze.js
@@ -35,7 +35,6 @@ const del_1 = __importDefault(require("del"));
 const yaml = __importStar(require("js-yaml"));
 const autobuild_1 = require("./autobuild");
 const codeql_1 = require("./codeql");
-const config_utils_1 = require("./config-utils");
 const diagnostics_1 = require("./diagnostics");
 const environment_1 = require("./environment");
 const feature_flags_1 = require("./feature-flags");
@@ -44,6 +43,7 @@ const tools_features_1 = require("./tools-features");
 const tracer_config_1 = require("./tracer-config");
 const upload_lib_1 = require("./upload-lib");
 const util = __importStar(require("./util"));
+const util_1 = require("./util");
 class CodeQLAnalysisError extends Error {
     constructor(queriesStatusReport, message) {
         super(message);
@@ -76,14 +76,14 @@ async function runExtraction(codeql, config, logger) {
             if (config.buildMode &&
                 (await codeql.supportsFeature(tools_features_1.ToolsFeature.TraceCommandUseBuildMode))) {
                 if (language === languages_1.Language.cpp &&
-                    config.buildMode === config_utils_1.BuildMode.Autobuild) {
+                    config.buildMode === util_1.BuildMode.Autobuild) {
                     await (0, autobuild_1.setupCppAutobuild)(codeql, logger);
                 }
                 try {
                     await codeql.extractUsingBuildMode(config, language);
                 }
                 catch (e) {
-                    if (config.buildMode === config_utils_1.BuildMode.Autobuild) {
+                    if (config.buildMode === util_1.BuildMode.Autobuild) {
                         const prefix = "We were unable to automatically build your code. " +
                             "Please change the build mode for this language to manual and specify build steps " +
                             "for your project. For more information, see " +
@@ -107,8 +107,8 @@ async function runExtraction(codeql, config, logger) {
 }
 exports.runExtraction = runExtraction;
 function shouldExtractLanguage(config, language) {
-    return (config.buildMode === config_utils_1.BuildMode.None ||
-        (config.buildMode === config_utils_1.BuildMode.Autobuild &&
+    return (config.buildMode === util_1.BuildMode.None ||
+        (config.buildMode === util_1.BuildMode.Autobuild &&
             process.env[environment_1.EnvVar.AUTOBUILD_DID_COMPLETE_SUCCESSFULLY] !== "true") ||
         (!config.buildMode && (0, languages_1.isScannedLanguage)(language)));
 }
