+++ lib/autobuild.js
@@ -28,7 +28,6 @@ const core = __importStar(require("@actions/core"));
 const actions_util_1 = require("./actions-util");
 const api_client_1 = require("./api-client");
 const codeql_1 = require("./codeql");
-const config_utils_1 = require("./config-utils");
 const environment_1 = require("./environment");
 const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
@@ -36,9 +35,9 @@ const repository_1 = require("./repository");
 const tools_features_1 = require("./tools-features");
 const util_1 = require("./util");
 async function determineAutobuildLanguages(codeql, config, logger) {
-    if ((config.buildMode === config_utils_1.BuildMode.None &&
+    if ((config.buildMode === util_1.BuildMode.None &&
         (await codeql.supportsFeature(tools_features_1.ToolsFeature.TraceCommandUseBuildMode))) ||
-        config.buildMode === config_utils_1.BuildMode.Manual) {
+        config.buildMode === util_1.BuildMode.Manual) {
         logger.info(`Using ${config.buildMode} build mode, nothing to autobuild.`);
         return undefined;
     }
