+++ lib/config-utils.test.js
@@ -35,7 +35,6 @@ const sinon = __importStar(require("sinon"));
 const api = __importStar(require("./api-client"));
 const codeql_1 = require("./codeql");
 const configUtils = __importStar(require("./config-utils"));
-const config_utils_1 = require("./config-utils");
 const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
@@ -269,7 +268,7 @@ function mockListLanguages(languages) {
         // And the config we expect it to parse to
         const expectedConfig = {
             languages: [languages_1.Language.javascript],
-            buildMode: config_utils_1.BuildMode.None,
+            buildMode: util_1.BuildMode.None,
             originalUserInput: {
                 name: "my config",
                 "disable-default-queries": true,
@@ -768,19 +767,19 @@ const mockRepositoryNwo = (0, repository_1.parseRepositoryNwo)("owner/repo");
 (0, ava_1.default)("Build mode not overridden when disable Java buildless feature flag disabled", async (t) => {
     const messages = [];
     const buildMode = await configUtils.parseBuildModeInput("none", [languages_1.Language.java], (0, testing_utils_1.createFeatures)([]), (0, testing_utils_1.getRecordingLogger)(messages));
-    t.is(buildMode, config_utils_1.BuildMode.None);
+    t.is(buildMode, util_1.BuildMode.None);
     t.deepEqual(messages, []);
 });
 (0, ava_1.default)("Build mode not overridden for other languages", async (t) => {
     const messages = [];
     const buildMode = await configUtils.parseBuildModeInput("none", [languages_1.Language.python], (0, testing_utils_1.createFeatures)([feature_flags_1.Feature.DisableJavaBuildlessEnabled]), (0, testing_utils_1.getRecordingLogger)(messages));
-    t.is(buildMode, config_utils_1.BuildMode.None);
+    t.is(buildMode, util_1.BuildMode.None);
     t.deepEqual(messages, []);
 });
 (0, ava_1.default)("Build mode overridden when analyzing Java and disable Java buildless feature flag enabled", async (t) => {
     const messages = [];
     const buildMode = await configUtils.parseBuildModeInput("none", [languages_1.Language.java], (0, testing_utils_1.createFeatures)([feature_flags_1.Feature.DisableJavaBuildlessEnabled]), (0, testing_utils_1.getRecordingLogger)(messages));
-    t.is(buildMode, config_utils_1.BuildMode.Autobuild);
+    t.is(buildMode, util_1.BuildMode.Autobuild);
     t.deepEqual(messages, [
         {
             message: "Scanning Java code without a build is temporarily unavailable. Falling back to 'autobuild' build mode.",
