+++ lib/status-report.test.js
@@ -29,7 +29,6 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const ava_1 = __importDefault(require("ava"));
 const sinon = __importStar(require("sinon"));
 const actionsUtil = __importStar(require("./actions-util"));
-const config_utils_1 = require("./config-utils");
 const environment_1 = require("./environment");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
@@ -56,7 +55,7 @@ function setupEnvironmentAndStub(tmpDir) {
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         setupEnvironmentAndStub(tmpDir);
         const statusReport = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.Init, "failure", new Date("May 19, 2023 05:19:00"), (0, testing_utils_1.createTestConfig)({
-            buildMode: config_utils_1.BuildMode.None,
+            buildMode: util_1.BuildMode.None,
             languages: [languages_1.Language.java, languages_1.Language.swift],
         }), { numAvailableBytes: 100, numTotalBytes: 500 }, (0, logging_1.getRunnerLogger)(false), "failure cause", "exception stack trace");
         t.truthy(statusReport);
@@ -67,7 +66,7 @@ function setupEnvironmentAndStub(tmpDir) {
             t.is(statusReport.action_started_at, new Date("May 19, 2023 05:19:00").toISOString());
             t.is(statusReport.actions_event_name, "dynamic");
             t.is(statusReport.analysis_key, "analysis-key");
-            t.is(statusReport.build_mode, config_utils_1.BuildMode.None);
+            t.is(statusReport.build_mode, util_1.BuildMode.None);
             t.is(statusReport.cause, "failure cause");
             t.is(statusReport.commit_oid, process.env["GITHUB_SHA"]);
             t.is(statusReport.exception, "exception stack trace");
